Replace long long with qint64
We are using it instead of int64_t because is plays nicer with QVariant.
Also add helper to use QJsonValue::toInteger with Qt 6.

+++ peer.h
@@ -23,8 +23,8 @@ namespace libtremotesf {
 
         QString address{};
         QString client{};
-        long long downloadSpeed{};
-        long long uploadSpeed{};
+        qint64 downloadSpeed{};
+        qint64 uploadSpeed{};
         double progress{};
         QString flags{};
     };
